Extract active-engine registry into engine_registry.py (WS5 seam)
Move the process-wide active-LCM-clone registry out of engine.py into a new
engine_registry.py module: the registry lock, the two WeakValueDictionary
registries (by session id / conversation id), the resolver/matcher helpers
(_is_usable_lcm_engine, _engine_matches_session_binding,
_engine_matches_conversation_binding), the removal helper
(_remove_registry_entries_for_engine), and the public resolve_active_lcm_engine.

engine.py imports the shared lock, the two registries, the removal helper, and
resolve_active_lcm_engine. The LCMEngine binding methods
(_register_active_engine_binding / _unregister_active_engine_binding) keep their
bodies verbatim: the lock and the weak registries are shared mutable objects, so
mutating them through the imported names hits the same objects. The matcher
helpers are internal to the module. resolve_active_lcm_engine is re-exported from
engine.py so hosts importing it from hermes_lcm.engine keep working.

Behaviour-preserving refactor (WS5 engine decomposition). The moved code is
byte-identical to its pre-move form. Rollback: git revert this commit restores
the inline block; no data or state migration involved.
